+++ src/Dto/PlaylistSlide.php
@@ -5,11 +5,13 @@ declare(strict_types=1);
 namespace App\Dto;
 
 use App\Dto\Trait\IdentifiableTrait;
+use App\Dto\Trait\RelationsModifiedTrait;
 use Symfony\Component\Serializer\Annotation\Groups;
 
 class PlaylistSlide
 {
     use IdentifiableTrait;
+    use RelationsModifiedTrait;
 
     #[Groups(['playlist-slide:read', 'slides/playlists:read'])]
     public Slide $slide;