+++ src/Dto/ScreenCampaign.php
@@ -6,6 +6,7 @@ namespace App\Dto;
 
 use App\Dto\Trait\BlameableTrait;
 use App\Dto\Trait\IdentifiableTrait;
+use App\Dto\Trait\RelationsModifiedTrait;
 use App\Dto\Trait\TimestampableTrait;
 use Symfony\Component\Serializer\Annotation\Groups;
 
@@ -14,6 +15,7 @@ class ScreenCampaign
     use BlameableTrait;
     use IdentifiableTrait;
     use TimestampableTrait;
+    use RelationsModifiedTrait;
 
     #[Groups(['screen-campaigns:read', 'campaigns/screens:read'])]
     public Playlist $campaign;