+++ src/Entity/Interfaces/TimestampableInterface.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Entity\Interfaces;
+
+interface TimestampableInterface
+{
+    public function getCreatedAt(): \DateTimeInterface;
+
+    public function setCreatedAt(): self;
+
+    public function getModifiedAt(): ?\DateTimeInterface;
+
+    public function setModifiedAt(): self;
+}