+++ src/State/ScreenGroupProvider.php
@@ -36,6 +36,8 @@ class ScreenGroupProvider extends AbstractProvider
         $output->campaigns = $iri.'/campaigns';
         $output->screens = $iri.'/screens';
 
+        $output->setRelationsModified($object->getRelationsModified());
+
         return $output;
     }
 }